+++ cli_folder/cli.py
@@ -1,4 +1,3 @@
-import game
 from game.chess import Chess
 
 def main():