+++ test/test_cli.py
@@ -48,10 +48,10 @@ class TestCli(unittest.TestCase):
     @patch('builtins.input', side_effect=['1', 'A', '2', 'A', 'y'],)
     @patch('builtins.print')
     @patch.object(Chess, 'move_piece_board')
-    def test_play_works(self, mock_chess_move, mock_print, mock_input,):
+    def test_play_works(self, mock_chess_move, mock_print, mock_input):
         main()
         self.assertEqual(mock_input.call_count, 5)
-        self.assertEqual(mock_print.call_count, 20)
+        self.assertEqual(mock_print.call_count, 22)
         self.assertEqual(mock_chess_move.call_count, 1)
 
 if __name__ == '__main__':